+++ packages/sdk-governance/src/bounties/bounties-sdk.ts
@@ -11,8 +11,11 @@ import {
 } from "./find-referenda"
 import { scheduledFinder } from "./find-scheduled"
 import { BountiesSdk, Bounty, GenericBounty, ProposedBounty } from "./sdk-types"
+import { PolkadotRuntimeOriginCaller } from "@/referenda/descriptors"
 
-export function createBountiesSdk(typedApi: BountiesSdkTypedApi): BountiesSdk {
+export function createBountiesSdk<TOrigin extends PolkadotRuntimeOriginCaller>(
+  typedApi: BountiesSdkTypedApi<TOrigin>,
+): BountiesSdk<{ origin: TOrigin }> {
   const { findScheduledApproved, findScheduledCuratorProposed } =
     scheduledFinder(typedApi)
 
@@ -20,7 +23,7 @@ export function createBountiesSdk(typedApi: BountiesSdkTypedApi): BountiesSdk {
     bounty: BountyWithoutDescription,
     description: string | null,
     id: number,
-  ): Bounty => {
+  ): Bounty<{ origin: TOrigin }> => {
     const generic: GenericBounty = {
       ...bounty,
       type: bounty.status.type,
@@ -48,7 +51,8 @@ export function createBountiesSdk(typedApi: BountiesSdkTypedApi): BountiesSdk {
           },
         }
       case "Approved":
-        return { ...generic, type: "Approved" }
+      case "ApprovedWithCurator":
+        return { ...generic, type: generic.status.type }
       case "Funded":
         return {
           ...generic,
@@ -180,7 +184,7 @@ export function createBountiesSdk(typedApi: BountiesSdkTypedApi): BountiesSdk {
 
   async function getProposedBounty(
     txEvent: TxEvent,
-  ): Promise<ProposedBounty | null> {
+  ): Promise<ProposedBounty<{ origin: TOrigin }> | null> {
     if (!("events" in txEvent)) {
       return null
     }