+++ packages/sdk-governance/src/bounties/find-referenda.ts
@@ -1,6 +1,10 @@
 import { OngoingReferendum } from "@/referenda/sdk-types"
 import { keyedMemo } from "@/util/memo"
 import { MultiAddress } from "./descriptors"
+import {
+  PolkadotRuntimeOriginCaller,
+  PolkadotRuntimeOriginCallerOriginal,
+} from "@/referenda/descriptors"
 
 const spenderOrigins = [
   "Treasurer",
@@ -11,35 +15,42 @@ const spenderOrigins = [
   "BigTipper",
 ]
 
-const getDecodedSpenderReferenda = keyedMemo(
-  async (ongoingReferenda: OngoingReferendum[]) => {
-    const spenderReferenda = ongoingReferenda.filter(
-      (ref) =>
-        (ref.origin.type === "Origins" &&
-          spenderOrigins.includes(ref.origin.value.type)) ||
-        (ref.origin.type === "system" && ref.origin.value.type === "Root"),
+const uncachedGetDecodedSpenderReferenda = async <
+  TOrigin extends PolkadotRuntimeOriginCaller,
+>(
+  ongoingReferenda: OngoingReferendum<{ origin: TOrigin }>[],
+) => {
+  const spenderReferenda = ongoingReferenda.filter((ref) => {
+    const origin = ref.origin as PolkadotRuntimeOriginCallerOriginal
+    return (
+      (origin.type === "Origins" &&
+        spenderOrigins.includes(origin.value.type)) ||
+      (origin.type === "system" && origin.value.type === "Root")
     )
-    const response = await Promise.all(
-      spenderReferenda.map((referendum) =>
-        referendum.proposal
-          .decodedCall()
-          .then((call) => ({
-            referendum,
-            call,
-          }))
-          .catch((ex) => {
-            console.error(ex)
-            return null
-          }),
-      ),
-    )
-    return response.filter((v) => !!v)
-  },
-  new WeakMap(),
-)
+  })
+  const response = await Promise.all(
+    spenderReferenda.map((referendum) =>
+      referendum.proposal
+        .decodedCall()
+        .then((call) => ({
+          referendum,
+          call,
+        }))
+        .catch((ex) => {
+          console.error(ex)
+          return null
+        }),
+    ),
+  )
+  return response.filter((v) => !!v)
+}
+const getDecodedSpenderReferenda: typeof uncachedGetDecodedSpenderReferenda =
+  keyedMemo(uncachedGetDecodedSpenderReferenda, new WeakMap())
 
-export async function findApprovingReferenda(
-  ongoingReferenda: OngoingReferendum[],
+export async function findApprovingReferenda<
+  TOrigin extends PolkadotRuntimeOriginCaller,
+>(
+  ongoingReferenda: OngoingReferendum<{ origin: TOrigin }>[],
   bountyId: number,
 ) {
   const spenderReferenda = await getDecodedSpenderReferenda(ongoingReferenda)
@@ -57,8 +68,10 @@ export async function findApprovingReferenda(
     .map(({ referendum }) => referendum)
 }
 
-export async function findProposingCuratorReferenda(
-  ongoingReferenda: OngoingReferendum[],
+export async function findProposingCuratorReferenda<
+  TOrigin extends PolkadotRuntimeOriginCaller,
+>(
+  ongoingReferenda: OngoingReferendum<{ origin: TOrigin }>[],
   bountyId: number,
 ) {
   const spenderReferenda = await getDecodedSpenderReferenda(ongoingReferenda)